:memo: Updata Code Generalization
