+++ Generalization/Learning_Curve/Noiseless/3_linear_plot.py
@@ -3,10 +3,10 @@ import matplotlib.pyplot as plt
 from numpy.linalg import pinv
 
 def problem_data(X):
-    return np.sin(np.dot(np.pi, X))
+    return np.sin(np.pi * X)
 
 def init_theta(shape):
-    theta = np.array(np.zeros(shape))
+    theta = np.zeros(shape)
     return theta
 
 def cost_function(n, Y, Y_pred):
@@ -21,28 +21,27 @@ def normal_equation(X, Y):
     return theta
 
 def cal_mean_model(Ed_model):
-    # mean_theta = np.array([np.mean(all_theta[:, 0]), np.mean(all_theta[:, 1])])
-    mean_model = np.mean(Ed_model, axis=0)
+    mean_model = np.mean(Ed_model, axis=1)
     return mean_model
 
 if __name__ == "__main__":
-    X = np.linspace(-1, 1)
+    X = np.linspace(-1, 1)  # Use 100 points between -1 and 1
 
-    X_b = np.c_[np.zeros((len(X), 1)), X]
+    X_b = np.c_[np.ones((len(X), 1)), X]
     
     y = problem_data(X)
+    steps = 1000
 
-    steps = 10000
-
-    # ed_prediction_list = list()
-    
     training_sizes = [2, 4, 6, 8, 10, 12, 14, 16, 18, 20]
     
     list_Ein = []
     list_Eout = []
+    func_g = []
+    bias_list = []
 
     for m in training_sizes:
         Ein_steps = []
+        Eout_steps = []
         ed_prediction_list = []
         
         for _ in range(steps):
@@ -55,26 +54,28 @@ if __name__ == "__main__":
             normal_theta = normal_equation(X_b_sample, y_sample)
             theta_arr = np.array(normal_theta)
 
-            # Find y_pred to cal Eout
-            sample_linear = linear_model(X_b, theta_arr)
-            ed_prediction_list.append(sample_linear)
-
             # Find Cost for Ein
             train_linear = linear_model(X_b_sample, theta_arr)
             e = cost_function(m, y_sample, train_linear)
             Ein_steps.append(e)
             
+            # Find y_pred to cal Eout
+            normal_theta_val = normal_equation(X_b, y)
+            theta_arr_val = np.array(normal_theta_val)
+            sample_linear = linear_model(X_b, theta_arr_val)
+            val_e = cost_function(m, y, sample_linear)
+            Eout_steps.append(val_e)
 
-        ed_arr_prediction = np.array(ed_prediction_list)
-        mean_model = cal_mean_model(ed_arr_prediction)
-        bias_square = np.mean(np.square(mean_model - y))
-        var_x = np.mean(np.square(ed_arr_prediction - mean_model))
-        variance = np.mean(var_x)
-        E_out = bias_square + variance
+        mean_model = cal_mean_model(ed_prediction_list)
+        func_g.append(mean_model) 
+        # print(np.array(func_g))
+        list_Ein.append(np.mean(Ein_steps))
+        list_Eout.append(np.mean(Eout_steps))
 
-        list_Ein.append(np.mean(np.array([Ein_steps])))
-        list_Eout.append(E_out)
+        bias = np.mean(np.square(np.mean(func_g) - y))
+        bias_list.append(bias) 
 
+    plt.plot(training_sizes, np.array(bias_list), label=f'Bias: {np.mean(np.array(bias_list)):.2f}', linestyle='--')
     plt.plot(training_sizes, np.array(list_Ein), label='Training Error (Ein)', c='r')
     plt.plot(training_sizes, np.array(list_Eout), label='Validation Error (Eout)', c='b')
     plt.xlabel('Training Set Size')
