+++ Generalization/Learning_Curve/Noisy/3_linear_Origin_with_noise.py
